4/9/2018

+++ Engine/Core/Math/Vector2D.cpp
@@ -1,6 +1,6 @@
 #include "Vector2D.h"
 
-Vector2D::Vector2D(double i_x, double i_y) {
+Vector2D::Vector2D(float i_x, float i_y) {
 	x = i_x;
 	y = i_y;
 }
